t0066: refactor dir-iterator tests
Be consistent throughout the dir-iterator tests regarding the order in
which we:
  * create test directories
  * create expected outputs
  * test if actual and expected outputs differ

These changes improve the readability of dir-iterator tests, separating
setup tests as we want to test different iteration schemes over the same
hierarchy. They also simplify the introduction of new tests that cover
current and future iteration schemes by providing a standardized
structure for testing dir-iterator with various directory hierarchies.

Signed-off-by: Plato Kiorpelidis <kioplato@gmail.com>