t0066: rename subtest descriptions
Improve the structure of dir-iterator tests by renaming subtest
descriptions. This makes the introduction of new tests for future
iteration schemes easier, as they will share similar naming, only
differentiating over the iteration scheme part of the description
that the subtest is testing for.

These structured subtest descriptions also help when running the tests.
It's easier to understand what's tested and in case a test fails find
out which iteration scheme failed.

Signed-off-by: Plato Kiorpelidis <kioplato@gmail.com>